Header를 Protocol의 Inner Class로 바꾸고 프로토콜들 통폐합함
