--- DMSClass/src/protocol/EventProtocolHeader.java
@@ -1,8 +1,0 @@
-package protocol;
-
-class EventProtocolHeader extends AbstractSplittableHeader {
-
-    EventProtocolHeader(short length, byte type, byte direction, byte code) {
-        super(length, type, direction, code);
-    }
-}
