--- DMSClass/src/protocol/FileProtocolHeader.java
@@ -1,8 +1,0 @@
-package protocol;
-
-class FileProtocolHeader extends AbstractSplittableHeader {
-
-    FileProtocolHeader(short length, byte type, byte direction, byte code) {
-        super(length, type, direction, code);
-    }
-}
