+++ DMSClass/src/protocol/Protocol.java
@@ -6,6 +6,8 @@ import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.io.Serializable;
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
 import java.util.Arrays;
 
 // 프로토콜 쓰는법 (계획)
@@ -49,13 +51,90 @@ import java.util.Arrays;
 // 자바에선 생성자에서 예외가 던져지네... 문제가 없을까?
 
 public class Protocol {
-    public final AbstractHeader header;
-    private final byte[] body_bytes;
+    class Header {
+        public final short length;      // 2바이트, 전체 프로토콜 길이
+        public final ProtocolType type; // 1바이트, 프로토콜 타입
+        public final byte direction;    // 1바이트, 프로토콜 응답 방향
+        public final byte code;         // 2바이트, 프로토콜 코드
+        // 아래 3개는 body가 커져서 프로토콜 분리시 쓰임
+        public final boolean isSplitted;// 1바이트, 프로토콜 분리 여부
+        public final boolean isLast;    // 1바이트, 마지막 프로토콜인지 여부
+        public final short sequence;    // 2바이트, 시퀀스 넘버
+        // body 시작 인덱스를 알기 위해 자신의 길이를 가지고 있다.
+        public static final int header_length = 10;
+    
+        Header(short length, ProtocolType type, byte direction, byte code) {
+            this.length = length;
+            this.type = type;
+            this.direction = direction;
+            this.code = code;
+            this.header_length = this.getHeaderLength();
+        }
+    
+        // 받은 패킷으로부터 프로토콜 헤더 추출
+        static Header create(byte[] packet) {
+            if (packet.length < 5)
+                return null; // 먼가 짧은게 들어왔다!!
+    
+            // 아래는 공통적으로 쓰이는 헤더 부분
+            ByteBuffer bb = ByteBuffer.allocate(2);
+            bb.order(ByteOrder.LITTLE_ENDIAN);
+            bb.put(packet[0]);
+            bb.put(packet[1]);
+            short length = bb.getShort();
+            // 자바는 byte[]의 길이를 바로 알아낼수 있어서 여기서 한번 더 체크해봄
+            if (length != packet.length)
+                return null; // 먼가 이상하다!
+            ProtocolType type = ProtocolType.getType(packet[2]);
+            byte direction = packet[3];
+            byte code = packet[4];
+    
+            return create(length, type, direction, code);
+        }
+    
+        // 새로 헤더 하나 만듬
+        static Header create(short length, ProtocolType type, byte direction, byte code) {
+            switch (type) {
+            case UNDEFINED:
+                // 이 경우로 프로토콜이 생성되는 경우는 없음!
+                break;
+            case LOGIN:
+                // TODO : 구현해야함
+            case FILE:
+                // TODO : 구현해야함
+            case EVENT:
+                // TODO : 구현해야함
+            default:
+                // 이 경우로 프로토콜이 생성되면 안됨!
+                break;
+            }
+    
+            // 먼가 문제 생겨서 switch 내에서 return 안되면 null 리턴 == 예외다!
+            // 따라서 null 체크 하셈ㅋ or 예외 던지게 코드 수정
+            return null;
+        }
+    
+        protected int getHeaderLength(){
+            return 5;
+        }
+    
+        byte[] getBytes() {
+            ByteArrayOutputStream baos = new ByteArrayOutputStream();
+            baos.write(length);
+            baos.write(type.ordinal());
+            baos.write(direction);
+            baos.write(code);
+            return baos.toByteArray();
+        }
+    }
+
+    public final Header header;
     public final Serializable body;
+    protected final byte[] body_bytes;
 
     // 받은 패킷으로부터 새 프로토콜 만들고 head 할당함.
     public Protocol(byte[] packet) throws IOException, ClassNotFoundException {
-        this.header = AbstractHeader.create(packet);
+        this.header = Header.create(packet);
         this.body_bytes = Arrays.copyOfRange(packet, header.header_length, packet.length);
         this.body = deserialization(this.body_bytes);
     }
@@ -64,7 +143,7 @@ public class Protocol {
     public Protocol(short length, ProtocolType type, byte direction, byte code, Serializable obj) throws IOException {
         this.body_bytes = serialization(obj);
         this.body = obj;
-        this.header = AbstractHeader.create(length, (byte)type.ordinal(), direction, code);
+        this.header = Header.create(length, type, direction, code);
     }
 
     // head를 바이트 배열로 바꿔서 이를 body랑 합쳐 반환함.
@@ -74,10 +153,6 @@ public class Protocol {
         System.arraycopy(headbyte, 0, packet, 0, headbyte.length);
         System.arraycopy(body_bytes, 0, packet, headbyte.length, body_bytes.length);
         return packet;
-    }
-
-    public AbstractHeader getHeader() {
-        return header;
     }
 
     private static byte[] serialization(Serializable obj) throws IOException {
