+++ DMSClass/src/protocol/ProtocolCode.java
@@ -1,0 +1,5 @@
+package protocol;
+
+public enum ProtocolCode{
+    
+}
