+++ DMSClass/src/protocol/ProtocolDirection.java
@@ -1,0 +1,5 @@
+package protocol;
+
+public enum ProtocolDirection{
+    TO_SERVER, TO_CLIENT;
+}
